Update README.md
Hi, 

Can I add two article in this great repository. I have pull the request.

NextItNet is a pioneering work that introduces the autoregressive-based next-token prediction concept and proposed a generative architecture  for sequential recommendation. 

GRec is the esstentially BART4Rec, which share high similarity with the famous BART model but is earlier than BART in terms of its arxiv version.